requirements.txt
dotenv==0.9.9
google-genai==1.7.0
chromadb==1.0.0
langchain==0.3.23
langchain-google-genai==2.1.2